Fixed typo and return data of do_deploy

+++ 2-do_deploy_web_static.py
@@ -30,7 +30,7 @@ def do_pack():
 def do_deploy(archive_path):
     """ Do Deploy """
     # check the file if exits
-    if not os.path.exist(archive_path):
+    if not os.path.exists(archive_path):
         return False
 
     # web_static_20170315003959.tgz
@@ -38,19 +38,37 @@ def do_deploy(archive_path):
     # web_static_20170315003959
     file_name_wo_ext = os.path.splitext(file_name)[0]
 
-    put(archive_path, '/tmp/%s'.format(file_name))
-    run("mkdir -p /data/web_static/releases/%s".format(file_name_wo_ext))
+    rs = put(archive_path, '/tmp/%s'.format(file_name))
+    if rs.failed:
+        return False
+    rs = run("mkdir -p /data/web_static/releases/%s".format(file_name_wo_ext))
+    if rs.failed:
+        return False
 
-    run("tar -xzf /tmp/%s -C /data/web_static/releases/%s/".format(
+    rs = run("tar -xzf /tmp/%s -C /data/web_static/releases/%s/".format(
         file_name, file_name_wo_ext))
+    if rs.failed:
+        return False
     
-    run("rm /tmp/%s".format(file_name))
-    run("mv /data/web_static/releases/%s/web_static/* \
+    rs = run("rm /tmp/%s".format(file_name))
+    if rs.failed:
+        return False
+    rs = run("mv /data/web_static/releases/%s/web_static/* \
         /data/web_static/releases/%s/".format(file_name_wo_ext,
                                               file_name_wo_ext))
-    run("rm -rf /data/web_static/releases/%s/web_static".format(
+    if rs.failed:
+        return False
+    rs = run("rm -rf /data/web_static/releases/%s/web_static".format(
         file_name_wo_ext))
+    if rs.failed:
+        return False
 
-    run("rm -rf /data/web_static/current")
-    run("ln -s /data/web_static/releases/%s/ /data/web_static/current".format(
-        file_name_wo_ext))
+    rs = run("rm -rf /data/web_static/current")
+    if rs.failed:
+        return False
+    rs = run("ln -s /data/web_static/releases/%s/ \
+             /data/web_static/current".format(file_name_wo_ext))
+    if rs.failed:
+        return False
+
+    return True
